console: Remove old dropdown
