--- pkg/webui/components/dropdown-v2/dropdown.styl
@@ -1,134 +1,0 @@
-// Copyright © 2023 The Things Network Foundation, The Things Industries B.V.
-//
-// Licensed under the Apache License, Version 2.0 (the "License");
-// you may not use this file except in compliance with the License.
-// You may obtain a copy of the License at
-//
-//     http://www.apache.org/licenses/LICENSE-2.0
-//
-// Unless required by applicable law or agreed to in writing, software
-// distributed under the License is distributed on an "AS IS" BASIS,
-// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-// See the License for the specific language governing permissions and
-// limitations under the License.
-
-ul.dropdown
-  font-size: $fs.m
-  border-radius: $br.l
-  border: 1px solid $c.grey-200
-  list-style-type: none
-  background: $c.white
-  box-shadow: 0px 3px 16px 0px rgba(0, 0, 0, 0.06)
-  position: absolute
-  padding: $cs.s
-  min-width: 14rem
-  z-index: $zi.dropdown
-  margin: $cs.xxs 0 0
-  opacity: 0
-  visibility: hidden
-  transition: opacity 0.2s ease-in-out, visibility 0s linear 0.2s
-
-  &.open
-    opacity: 1
-    visibility: visible
-    transition: opacity 0.2s ease-in-out, visibility 0s linear
-
-  &.below
-    bottom: auto
-    top: calc(100% + 7px)
-
-  &.above
-    top: auto
-    bottom: calc(100% + 7px)
-
-  &.left
-    left: 0
-    right: auto
-
-  &.right
-    right: 0
-    left: auto
-
-  &.example
-    bottom: auto
-    top: 0
-    left: 0
-    right: auto
-
-  &.larger
-    li.dropdown-item
-      & > a.button, & > button.button
-        padding: $cs.l
-
-  hr
-    height: 1px
-    background-color: $c.grey-200
-
-  li.dropdown-header-item
-    display: block
-    margin-bottom: 0
-    font-weight: $fw.bold
-
-    span
-      line-height: 1
-      display: block
-      padding: $cs.m
-
-  li.dropdown-item
-    margin-bottom: 0
-    text-align: left
-
-    button.button
-      reset-button()
-
-    & > a.button, & > button.button
-      box-sizing: border-box
-      line-height: 1
-      display: flex
-      align-items: center
-      text-decoration: none
-      padding: $cs.xs
-      width: 100%
-      min-height: 2.429rem
-      &:not(.active)
-        color: $c.grey-700
-
-        .icon
-          color: $c.grey-700
-
-      &.active
-        border-radius: $br.m
-        background: $c.tts-primary-150
-        transition: background 0.1s ease-in-out
-        color: $tc-deep-gray
-
-        .icon
-          color: $tc-deep-gray
-
-      &:hover
-        color: $tc-deep-gray
-        background: $c.tts-primary-150
-        transition: background 0.1s ease-in-out
-        border-radius: $br.m
-
-        .icon
-          color: $tc-deep-gray
-
-    .icon
-      margin-right: $cs.xs
-
-    &:hover
-      color: $tc-deep-gray
-      text-decoration: none
-
-  .submenu-container
-    bottom: 0px
-    left: 0px
-    position: absolute
-    width: 180%
-    height: 5rem
-    z-index: 9999
-
-  .submenu-dropdown
-    bottom: -0px !important
-    left: calc(100% + -213px) !important
