+++ pkg/webui/components/dropdown/index.js
@@ -1,4 +1,4 @@
-// Copyright © 2020 The Things Network Foundation, The Things Industries B.V.
+// Copyright © 2023 The Things Network Foundation, The Things Industries B.V.
 //
 // Licensed under the Apache License, Version 2.0 (the "License");
 // you may not use this file except in compliance with the License.
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-import React, { useCallback } from 'react'
+import React, { useCallback, useEffect, useRef, useState } from 'react'
 import classnames from 'classnames'
 import { NavLink } from 'react-router-dom'
 
@@ -25,21 +25,64 @@ import PropTypes from '@ttn-lw/lib/prop-types'
 
 import style from './dropdown.styl'
 
-const Dropdown = React.forwardRef(({ className, children, larger, onItemsClick }, ref) => (
-  <ul
-    onClick={onItemsClick}
-    className={classnames(style.dropdown, className, { [style.larger]: larger })}
-    ref={ref}
-  >
-    {children}
-  </ul>
-))
+const Dropdown = ({ className, children, larger, onItemsClick, open }) => {
+  const ref = useRef(null)
+  const [isBelow, setIsBelow] = useState(false)
+  const [isOnRight, setIsOnRight] = useState(false)
+
+  const positionDropdown = useCallback(() => {
+    if (ref.current) {
+      const parentRect = ref.current.parentElement.getBoundingClientRect()
+      const spaceBelow = window.innerHeight - parentRect.bottom
+      const spaceAbove = parentRect.top
+      const dropdownHeight = ref.current.clientHeight
+      const dropdownWidth = ref.current.clientWidth
+      const spaceOnLeft = parentRect.left
+
+      setIsBelow(spaceBelow > dropdownHeight || spaceAbove < dropdownHeight)
+
+      setIsOnRight(spaceOnLeft > dropdownWidth)
+    }
+  }, [])
+
+  useEffect(() => {
+    if (open) positionDropdown()
+  }, [positionDropdown, open])
+
+  useEffect(() => {
+    window.addEventListener('scroll', positionDropdown)
+    window.addEventListener('resize', positionDropdown)
+
+    return () => {
+      window.removeEventListener('resize', positionDropdown)
+      window.removeEventListener('scroll', positionDropdown)
+    }
+  }, [positionDropdown])
+
+  return (
+    <ul
+      onClick={onItemsClick}
+      className={classnames(style.dropdown, className, {
+        [style.larger]: larger,
+        [style.below]: isBelow,
+        [style.above]: !isBelow,
+        [style.right]: isOnRight,
+        [style.left]: !isOnRight,
+        [style.open]: open,
+      })}
+      ref={ref}
+    >
+      {children}
+    </ul>
+  )
+}
 
 Dropdown.propTypes = {
   children: PropTypes.node.isRequired,
   className: PropTypes.string,
   larger: PropTypes.bool,
   onItemsClick: PropTypes.func,
+  open: PropTypes.bool.isRequired,
 }
 
 Dropdown.defaultProps = {
@@ -58,8 +101,10 @@ const DropdownItem = ({
   showActive,
   tabIndex,
   external,
+  submenuItems,
   ...rest
 }) => {
+  const [expandedSubmenu, setExpandedSubmenu] = useState(false)
   const iconElement = icon && <Icon className={style.icon} icon={icon} nudgeUp />
   const activeClassName = classnames({
     [style.active]: (!Boolean(action) && showActive) || active,
@@ -90,9 +135,39 @@ const DropdownItem = ({
       <Message content={title} />
     </NavLink>
   )
+
+  const handleMouseEnter = useCallback(() => {
+    setExpandedSubmenu(true)
+  }, [setExpandedSubmenu])
+
+  const handleMouseLeave = useCallback(() => {
+    setExpandedSubmenu(false)
+  }, [setExpandedSubmenu])
+
+  const withSubmenu = (
+    <button
+      onMouseEnter={handleMouseEnter}
+      onMouseLeave={handleMouseLeave}
+      className={classnames(style.button, 'd-flex', 'j-between')}
+    >
+      <div>
+        {iconElement}
+        <Message content={title} />
+      </div>
+      <Icon icon="chevron_right" />
+      {expandedSubmenu ? (
+        <div className={style.submenuContainer}>
+          <Dropdown open={expandedSubmenu} className={style.submenuDropdown}>
+            {submenuItems}
+          </Dropdown>
+        </div>
+      ) : null}
+    </button>
+  )
+
   return (
     <li className={style.dropdownItem} key={title.id || title} {...rest}>
-      {ItemElement}
+      {Boolean(submenuItems) ? withSubmenu : ItemElement}
     </li>
   )
 }
@@ -105,6 +180,7 @@ DropdownItem.propTypes = {
   icon: PropTypes.string,
   path: PropTypes.string,
   showActive: PropTypes.bool,
+  submenuItems: PropTypes.arrayOf(PropTypes.node),
   tabIndex: PropTypes.string,
   title: PropTypes.message.isRequired,
 }
@@ -118,6 +194,7 @@ DropdownItem.defaultProps = {
   path: undefined,
   showActive: true,
   tabIndex: '0',
+  submenuItems: undefined,
 }
 
 const DropdownHeaderItem = ({ title }) => (
